requirements.txt
flask
flask_wtf
numpy
scipy
moviepy
pillow
wtforms
flask_debugtoolbar